feat: secure changing methods

+++ travel-app-api/src/controllers/country.controller.ts
@@ -1,35 +1,24 @@
-import {
-  Count,
-  CountSchema,
-  Filter,
-  FilterExcludingWhere,
-  repository,
-  Where,
-} from '@loopback/repository';
-import {
-  post,
-  param,
-  get,
-  getModelSchemaRef,
-  patch,
-  put,
-  del,
-  requestBody,
-  response,
-} from '@loopback/rest';
-import {Country, CountryWithRelations} from '../models';
-import {CountryRepository} from '../repositories';
+import { repository } from '@loopback/repository';
+import { post, param, get, getModelSchemaRef, patch, del, requestBody, response, HttpErrors } from '@loopback/rest';
+import { Country } from '../models';
+import { authenticate } from '@loopback/authentication';
+import { SecurityBindings, securityId, UserProfile } from '@loopback/security';
+import { UserRepository } from '@loopback/authentication-jwt';
+import { inject } from '@loopback/core';
+import { CountryRepository } from '../repositories';
 
 export class CountryController {
   constructor(
     @repository(CountryRepository)
-    public countryRepository : CountryRepository,
+    public countryRepository: CountryRepository,
+    @repository(UserRepository) protected userRepository: UserRepository,
   ) {}
 
+  @authenticate('jwt')
   @post('/countries')
   @response(200, {
     description: 'Country model instance',
-    content: {'application/json': {schema: getModelSchemaRef(Country)}},
+    content: { 'application/json': { schema: getModelSchemaRef(Country) } },
   })
   async create(
     @requestBody({
@@ -43,7 +32,13 @@ export class CountryController {
       },
     })
     country: Omit<Country, 'id'>,
+    @inject(SecurityBindings.USER)
+    currentUserProfile: UserProfile,
   ): Promise<Country> {
+    const user = await this.userRepository.findById(currentUserProfile[securityId]);
+    if (!user.isAdmin) {
+      throw new HttpErrors[403]();
+    }
     return this.countryRepository.create(country);
   }
 
@@ -54,7 +49,7 @@ export class CountryController {
       'application/json': {
         schema: {
           type: 'array',
-          items: getModelSchemaRef(Country, {includeRelations: true}),
+          items: getModelSchemaRef(Country, { includeRelations: true }),
         },
       },
     },
@@ -62,22 +57,21 @@ export class CountryController {
   async find(): Promise<Country[]> {
     return this.countryRepository.find();
   }
-  
+
   @get('/countries/{code}')
   @response(200, {
     description: 'Country model instance',
     content: {
       'application/json': {
-        schema: getModelSchemaRef(Country, {includeRelations: true}),
+        schema: getModelSchemaRef(Country, { includeRelations: true }),
       },
     },
   })
-  async findById(
-    @param.path.string('code') code: string,
-  ): Promise<Country | null> {
-    return this.countryRepository.findOne({ where: {code}, include: [{ relation: "sights" }] });
+  async findById(@param.path.string('code') code: string): Promise<Country | null> {
+    return this.countryRepository.findOne({ where: { code }, include: [{ relation: 'sights' }] });
   }
 
+  @authenticate('jwt')
   @patch('/countries/{id}')
   @response(204, {
     description: 'Country PATCH success',
@@ -87,20 +81,35 @@ export class CountryController {
     @requestBody({
       content: {
         'application/json': {
-          schema: getModelSchemaRef(Country, {partial: true}),
+          schema: getModelSchemaRef(Country, { partial: true }),
         },
       },
     })
     country: Country,
+    @inject(SecurityBindings.USER)
+    currentUserProfile: UserProfile,
   ): Promise<void> {
+    const user = await this.userRepository.findById(currentUserProfile[securityId]);
+    if (!user.isAdmin) {
+      throw new HttpErrors[403]();
+    }
     await this.countryRepository.updateById(id, country);
   }
 
+  @authenticate('jwt')
   @del('/countries/{id}')
   @response(204, {
     description: 'Country DELETE success',
   })
-  async deleteById(@param.path.string('id') id: string): Promise<void> {
+  async deleteById(
+    @param.path.string('id') id: string,
+    @inject(SecurityBindings.USER)
+    currentUserProfile: UserProfile,
+  ): Promise<void> {
+    const user = await this.userRepository.findById(currentUserProfile[securityId]);
+    if (!user.isAdmin) {
+      throw new HttpErrors[403]();
+    }
     await this.countryRepository.deleteById(id);
   }
 }
